Move package aggregation to the Source class
Given that we already have a class for package sources, it makes more sense for
that functionality to live in that class, instead of the manager.

This also introduces a small tweak to the behavior of the refresh command,
prohibiting the use of "--push" without "--aggregate". There is no other local
modification to source repo clones.
