+++ zeekpkg/package.py
@@ -3,12 +3,14 @@ A module with various data structures used for interacting with and querying
 the properties and status of Zeek packages.
 """
 
+import configparser
 import os
 import re
 from functools import total_ordering
 
 import semantic_version as semver
 
+from . import LOG
 from ._util import find_sentence_end, normalize_version_tag
 from .uservar import UserVar
 
@@ -28,6 +30,41 @@ PLUGIN_MAGIC_FILE = "__zeek_plugin__"
 PLUGIN_MAGIC_FILE_DISABLED = "__zeek_plugin__.disabled"
 LEGACY_PLUGIN_MAGIC_FILE = "__bro_plugin__"
 LEGACY_PLUGIN_MAGIC_FILE_DISABLED = "__bro_plugin__.disabled"
+
+
+def get_package_metadata(parser: configparser.ConfigParser) -> dict[str, str]:
+    return {item[0]: item[1] for item in parser.items("package")}
+
+
+def pick_metadata_file(directory: str) -> str:
+    rval = os.path.join(directory, METADATA_FILENAME)
+
+    if os.path.exists(rval):
+        return rval
+
+    return os.path.join(directory, LEGACY_METADATA_FILENAME)
+
+
+def parse_package_metadata(
+    parser: configparser.ConfigParser,
+    metadata_file: str,
+) -> str:
+    """Return string explaining why metadata is invalid, or '' if valid."""
+    if not parser.read(metadata_file):
+        LOG.warning("%s: missing metadata file", metadata_file)
+        return (
+            f"missing {METADATA_FILENAME} (or {LEGACY_METADATA_FILENAME}) metadata file"
+        )
+
+    if not parser.has_section("package"):
+        LOG.warning("%s: metadata missing [package]", metadata_file)
+        return f"{os.path.basename(metadata_file)} is missing [package] section"
+
+    for a in aliases(get_package_metadata(parser)):
+        if not is_valid_name(a):
+            return f'invalid alias "{a}"'
+
+    return ""
 
 
 def name_from_path(path: str) -> str:
